Refactor get balance functions to WalletApplication.kt
